[round-4] 동시성 처리 및 쿠폰 도메인 적용
* 주문시 포인트 차감, 재고 차감 동시성 처리 및 테스트 구현

* feat: 주문 요청 동시성 테스트

* feat: 주문 요청 요구사항 변경으로 인한 포인트 차감에 대한 락 수정

* Week4/feature/coupon 도메인 추가

* fix: 쿠폰 설계 추가

* feat: 쿠폰 목록 조회 기능 구현

* feat: 쿠폰 목록 조회 테스트 구현

* feat: 주문 쿠폰 적용

* feat: 좋아요 동시성 테스트 구현

+++ apps/commerce-api/src/main/java/com/loopers/application/point/PointCreateInfo.java
@@ -2,17 +2,19 @@ package com.loopers.application.point;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class PointCreateInfo {
     private Long userId;
-    private int point;
+    private BigDecimal point;
 
-    public PointCreateInfo(Long userId, int point) {
+    public PointCreateInfo(Long userId, BigDecimal point) {
         this.userId = userId;
         this.point = point;
     }
 
-    public static PointCreateInfo of(Long userId, int point) {
+    public static PointCreateInfo of(Long userId, BigDecimal point) {
         return new PointCreateInfo(userId, point);
     }
 }
